+++ CHANGES.md
@@ -1,5 +1,6 @@
 #### [unreleased]
 * skip mu-plugins and drop-ins
+* filter options to save to remove unchecked and VCS repos
 
 #### 2.5.4 / 2020-06-01
 * sanitizing, escaping & ignoring